+++ engine/platform/Vulkan/VKrenderer.h
@@ -60,8 +60,8 @@ namespace GfxRenderEngine
 
         VkRenderPass GetSwapChainRenderPass() const { return m_SwapChain->GetRenderPass(); }
         float GetAspectRatio() const { return m_SwapChain->ExtentAspectRatio(); }
-        float GetContextWidth() const { return m_SwapChain->Width(); }
-        float GetContextHeight() const { return m_SwapChain->Height(); }
+        uint GetContextWidth() const { return m_SwapChain->Width(); }
+        uint GetContextHeight() const { return m_SwapChain->Height(); }
         bool FrameInProgress() const { return m_FrameInProgress; }
 
         VkCommandBuffer GetCurrentCommandBuffer() const;
